TimetableServiceHandlersRegistrator.py renamed into timetable_service_handlers_registrar.py and moved into service_handlers_registrars
